requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
requests==2.32.3
urllib3==1.26.15
Werkzeug==3.0.4
azure-cognitiveservices-speech==1.30.0
azure-ai-formrecognizer==3.3.3  # Add the Azure Form Recognizer SDK
azure-core==1.26.3  # Add Azure Core SDK, a dependency for Azure SDKs
python-dotenv==0.19.2
waitress==2.1.2